% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{bamlssMVN-package}
\alias{bamlssMVN}
\alias{bamlssMVN-package}
\title{bamlssMVN: MVN Cholesky family for BAMLSS}
\description{
Implements a prototype for a BAMLSS family modelling a multi-variate
             Gaussian distribution via the Cholesky decomposition of its
             co-varaince matrix.
}
\details{
This is the package bamlss_mvn.
It implements bamlss families for modelling multivariate
responses. The variance-covariance matrix is re-parameterized by
the (modified) Cholesky decomposition.
The key function to generate a family is \code{mvnchol_bamlss}.
You can adjust the dimension \code{k} of the multivariate normal,
and its \cdoe{type} of Cholesky decompositions: \code{"basic"} or
\code{"modified"}.
}
\author{
\strong{Maintainer}: Thomas Muschinski \email{Thomas.Muschinkski@uibk.ac.at}

Authors:
\itemize{
  \item Thorsten Simon \email{Thorsten.Simon@uibk.ac.at} (\href{https://orcid.org/0000-0002-3778-7738}{ORCID})
}

}
\keyword{internal}
