% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnchol_bamlss.R
\name{mvnchol_bamlss}
\alias{mvnchol_bamlss}
\title{Cholesky MVN}
\usage{
mvnchol_bamlss(k, type = c("chol", "modified"), ...)
}
\arguments{
\item{k}{integer. The dimension of the multivariate distribution.}

\item{type}{character. Choose standard Cholesky "chol" or modified
Cholesky "modified".}

\item{...}{not used.}
}
\value{
a bamlss family.
}
\description{
BAMLSS Families for MVN with Cholesky Parameterization
}
\details{
BAMLSS families that models a multivariate Normal (Gaussian)
distribution by (modified) Cholesky decomposition of the covariance
matrix.
}
