% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{sim_mvn}
\alias{sim_mvn}
\title{MVN Simulator}
\usage{
sim_mvn(n, k = 3, type = "AR1", rho = 0.9, ...)
}
\arguments{
\item{n}{integer. Sample size.}

\item{k}{integer. Dimension of mvn distribution.}

\item{type}{character. Type of parameterization (see details).}

\item{rho}{numeric. Vector of parameters.}

\item{...}{not used.}
}
\description{
Simulate from MVN
}
\details{
This simulator generates synthtic data from a multivaraite normal
with zero mean and unit standard deviation.
Various parameterizations can be selected (such as "AR1", "AR2",
"AD1", "AD2").
}
