% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{TempIbk}
\alias{TempIbk}
\title{Temperature data.}
\format{
An object of class \code{data.frame} with 1798 rows and 17 columns.
}
\usage{
data(TempIbk)
}
\description{
Temperature Data for Innsbruck Airport
}
\details{
Numerical weather predictions (NWP) and observations of
2 meter temperature at Innsbruck Airport.
The observations from the SYNOP station 11120 cover 5 years from
The NWP data are derived from GEFS reforecasts (Hamill et al. 2013).
The data contain following variables:

\itemize{
  \item \code{init}: Time of initialization of the NWP model.
  \item \code{obs_*}: Observations for lead time \code{*}.
  \item \code{mean_ens_*}: NWP ensemble mean for lead time \code{*}.
  \item \code{logsd_ens_*}: NWP logarithm of ensemble standard deviation for lead time \code{*}.
  \item \code{yday}: Yearday.
}
}
\examples{
\dontrun{
## packages and data
library("bamlss")
library("mvnchol")
data("TempIbk", package = "mvnchol")

## five lead times
lead <- seq(192, 216, by = 6)

## set up formulas
f <- c(
  ## mu equations
  sprintf('obs_\%s ~ s(yday, bs = "cc") + s(yday, bs = "cc", by = mean_ens_\%s)', lead, lead),

  ## lambda diag equations
  sprintf('lamdiag\%s ~ s(yday, bs = "cc") + s(yday, bs = "cc", by = logsd_ens_\%s)', 1:5, lead),

  ## lambda off-diag equations
  sprintf('lambda\%s ~ s(yday, bs = "cc")', apply(combn(1:5, 2), 2, paste, collapse = ""))
)
f <- lapply(f, as.formula)

## multivariate normal family with basic Cholesky parameterization
fam <- mvnchol_bamlss(k = 5, type = "basic")

## fit boosting model
b <- bamlss(f, family = fam, data = TempIbk, optimizer = opt_boost, maxit = 1000, sampler = FALSE)
}

}
\references{
Hamill TM, Bates GT, Whitaker JS, Murray DR, Fiorino M, Galarneau Jr TJ,
    Zhu Y, Lapenta W (2013). NOAA's Second-Generation Global Medium-Range
    Ensemble Reforecast Data Set. \emph{Bulletin of the American Meteorological
    Society}, 94(10), 1553-1565.
}
\keyword{datasets}
