% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_formula.R
\name{make_formula}
\alias{make_formula}
\title{Formula Generator}
\usage{
make_formula(formula, type = "chol")
}
\arguments{
\item{formula}{formula. E.g. O | C | E | A | N ~ x | y | z}

\item{type}{character. Type of Cholesky decomposition.}
}
\description{
Formula Generator
}
