% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{TempIbk}
\alias{TempIbk}
\title{Temperature data.}
\format{
An object of class \code{data.frame} with 1798 rows and 17 columns.
}
\usage{
data(TempIbk)
}
\description{
Temperature Data for Innsbruck Airport
}
\details{
Numerical weather predictions (NWP) and observations of
2 meter temperature at Innsbruck Airport.
The observations from the SYNOP station 11120 cover 5 years from
The NWP data are derived from GEFS reforecasts (Hamill et al. 2013).
The data contain following variables:

\itemize{
  \item \code{init}: Time of initialization of the NWP model.
  \item \code{obs_*}: Observations for lead time \code{*}.
  \item \code{mean_ens_*}: NWP ensemble mean for lead time \code{*}.
  \item \code{logsd_ens_*}: NWP logarithm of ensemble standard deviation for lead time \code{*}.
  \item \code{yday}: Yearday.
}
}
\examples{
\dontrun{
## packages and data
library("bamlss")
library("mvnchol")
data("TempIbk", package = "mvnchol")

## five lead times
lead <- seq(192, 216, by = 6)

## set up formulas
f <- c(
  ## mu equations
  sprintf('obs_\%s ~ s(yday, bs = "cc") + s(yday, bs = "cc", by = mean_ens_\%s)', lead, lead),

  ## lambda diag equations
  sprintf('lamdiag\%s ~ s(yday, bs = "cc") + s(yday, bs = "cc", by = logsd_ens_\%s)', 1:5, lead),

  ## lambda off-diag equations
  sprintf('lambda\%s ~ s(yday, bs = "cc")', apply(combn(1:5, 2), 2, paste, collapse = ""))
)
f <- lapply(f, as.formula)

## multivariate normal family with basic Cholesky parameterization
fam <- mvnchol_bamlss(k = 5, type = "basic")

## fit boosting model
b <- bamlss(f, family = fam, data = TempIbk, optimizer = opt_boost, maxit = 1000, sampler = FALSE)

## predict sample case
nd <- subset(TempIbk, init == "2015-10-10")
fit <- predict(b, newdata = nd, type = "parameter", mstop = 5000)

## plot correlation matrix for GEFS initialization 2015-10-10
image(lead, lead, fam$correlation(fit)[[1]][5:1, ],
	 col = hcl.colors(12, "Blues 3", rev = TRUE), axes = FALSE,
	 xlab = "lead time in hours", ylab = "lead time in hours",
	 main = "Correlation matrix fitted for 2015-10-10")
axis(1, lead)
axis(2, lead, rev(lead)
box()

## plot means and standard deviations
stdev <- fam$stdev(fit)[[1]]
means <- fam$means(fit)[[1]]
lower <- means - stdev
upper <- means + stdev

plot(lead, means, type = 'b', cex = 2, lwd = 1, lty = 2, axes = FALSE,
	 ylim = c(min(lower), max(upper)),
	 ylab = expression("Temperature in " * degree * "C"),
	 xlab = "lead time in hours",
	 main = "Means +/- one st. dev. for 2015-10-10")
segments(lead, y0 = lower, y1 = upper)
axis(1, lead)
axis(2)
box()
}

}
\references{
Hamill TM, Bates GT, Whitaker JS, Murray DR, Fiorino M, Galarneau Jr TJ,
    Zhu Y, Lapenta W (2013). NOAA's Second-Generation Global Medium-Range
    Ensemble Reforecast Data Set. \emph{Bulletin of the American Meteorological
    Society}, 94(10), 1553-1565.
}
\keyword{datasets}
